Fix test cases for numpy 2.0
- Some test objects now use numpy types explicitly
- types now set in tested State method(s) so they are less likely to be changed
- see: RAD-527

+++ examples/rover/rover_model.py
@@ -382,8 +382,8 @@ class ControlState(State):
         left power
     """
 
-    rpower: float = 0.0
-    lpower: float = 0.0
+    rpower: np.float64 = 0.0
+    lpower: np.float64 = 0.0
 
 
 class Control(Flow):
